+++ CHANGELOG.md
@@ -5,7 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.5.2] - 2023-22-03
+## Unreleased - 2023-27-05
+
+### Added
+
+- `Scalar` and `InternalVector` traits to help genericity of built-in `ComputeMethods`.
+- `Tree`, `Node`, `Orthant`, `BoundingBox` structs used by `BarnesHut` compute methods.
+- `TreeData`, `BoundingBoxDivide`, `BarnesHutTree` traits for `BarnesHut` compute methods.
+
+### Changed
+
+- Built-in `ComputeMethods` generic using `Scalar` and `InternalVector` traits.
+- `tree` module and submodules now public.
+
+## [0.5.2] - 2023-16-05
 
 ### Changed